requirements.txt
# core
requests
pyyaml
docopt

# for running models
pyqt5
ipython
matplotlib
scipy
2to3